pyproject.toml
[project]
name = "ecoaction-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=0.19.0",
    "numpy>=2.2.6",
    "openai>=1.100.2",
    "pandas>=2.3.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "email-validator>=2.0.0",
    "jinja2>=3.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.7",
    "asyncpg>=0.28.0",
    "postgrest>=0.16.0",
    "gotrue>=1.3.0",
    "realtime>=1.0.0",
    "storage3>=0.7.0",
    "streamlit>=1.48.1",
    "ipykernel>=6.30.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "crewai>=0.159.0",
    "pytest>=8.4.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "ruff>=0.1.0",
    "bandit>=1.7.5",
    "safety>=2.3.0",
]
